회원 탈퇴, 회원 수정, 회원 정보보기 시 권한 검증 +(spring 자동 re-run)

+++ src/main/java/com/example/prj1be20231109/controller/MemberController.java
@@ -67,16 +67,37 @@ public class MemberController {
     }
 
     @GetMapping
-    public ResponseEntity<Member> view(String id) {
+    public ResponseEntity<Member> view(String id, @SessionAttribute(value = "login", required = false) Member login) {
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();  // 401
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
+
+
         Member member = service.getMember(id);
 
         return ResponseEntity.ok(member);
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();  // 401
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
+
         // TODO : 로그인 했는지? -> 안했으면 401
         // TODO : 자기 정보인지? -> 자기정보 아니면 403
+        System.out.println("id = " + id);
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -86,7 +107,17 @@ public class MemberController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
+    public ResponseEntity edit(@RequestBody Member member,
+                               @SessionAttribute(value = "login", required = false) Member login) {
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if (!service.hasAccess(member.getId(), login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
+
         // TODO: 로그인 했는지? 자기정보인지?
 
         if (service.update(member)) {
